feat: Delete and supply can used

+++ ims-apple/IMS Apple/Networking/Managers/ProductManager.swift
@@ -68,7 +68,7 @@ final class ProductManager {
         }
     }
     
-    func supplyProduct(id: Int, with stock: Int) async throws {
+    func supplyProduct(id: Int, with stock: Double) async throws {
         let parameters: [String: Any] = [
             "id": id,
             "stock": stock
